backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
stripe==7.8.0
aiofiles==23.2.1
pillow==10.4.0
redis==5.0.1
cloudinary==1.37.0
bcrypt==4.1.2
apscheduler==3.10.4
websockets==12.0

